#19 extract dogu diff functions into own file
By the name, `stateDiff` should contain integrating functionalities.
On the other hand, it mostly contained dogu diff functions. With
component diffs being developed and the upcoming config value diffs,
this file is no longer a viable place for all these.

This commit extracts all dogu diff functions into its own file.
