+++ pkg/domain/stateDiffComponent.go
