+++ pkg/domain/stateDiffComponent_test.go
