+++ pkg/domain/stateDiffDogu_test.go
